[UFP-509] frontend does not have Dockerfile
Problem
-------

A Dockerfile will ease the deployment of the frontend.

Solution
--------

Create a Dockerfile for the frontend.

Notes
-----

The node adapter for svelte was selected.  Other options may be
possible.

The Dockerfile is parameterized by the `VITE_BACKEND_URL` argument.
This avoid server-side rendering at the cost of requiring a build for
each target environment.