UFP-779 Update Bootstrap Condition
Bootstrap will now only run when dpm.bootstrap.data.enabled is set to true.  The datasets that are loaded can continue to be environment specific.  For example, if dpm.bootstrap.data.enabled is set to true that will cause Bootstrap to run and if the dev-data environment is enabled, Bootstrap would load data from sources like application-dev-data.yml.

Whether or not Bootstrap runs is a separate concern from which data, if any, Bootstrap should load.