XML escaping is incorrect
Problem
-------

Strings escaped for XML should use &lt;, &gt;, etc.

Escaping the subject name should not have special rules, that is, the
normal XML escaping rules apply so that an XML parser can unescape
it.  The unescaped string would need to be parsed which may process
additional escapes.

Solution
--------

Use the correct XML escapes and simplify handling of the subject name
when escaping it for the permissions document.